Refactor jenkins to use alb
This includes changing the jenkins configuration so that the slave
containers talk back to the master via its private ip and changing the
security groups to allow it.

Because we no longer need cloudfront, this has been removed along with
the lambda which updated the security groups rules to allow traffic from
cloudfront.