Replace github username and password with SSH key
This makes it easier to let Jenkins jobs to push changes back to github.

Jenkins clones the repo using the public URL, which you can only push
change to if you are using an SSH key for authentication. Pushing changes
with a password would require you to add the username and password to the
git URL.

Co-authored-by: Loki Holbourn <loki.holbourn@nationalarchives.gov.uk>