Update Jenkinsfile
This now does three new things.

It copies the jobs folder into the /tmp directory. This stops tar
complaining that the folder is being modified in case anyone is changing
the jobs

It also deletes the currently running backup
build from the folder. This stops the job showing as running when
Jenkins restores.

Finally, it sends a message to slack if the build fails so we know
if it's failing.