Stop using the secrets file
Originally, we were getting the secrets from a file in S3, populating
systems manager parameter store with them and then using parameter store
to set the secrets. This isn't that secure so I've cut this bit out and
assumed that the secrets in the parameter store will be set manually
like a lot of the other ones are.