Remove /var/jenkins_home mount from Jenkins container
This directory contains Jenkins dependencies like plugins, build definitions
and build histories.

This directory was originally mounted from the EC2 instance so that we could
preserve its contents between ECS container deployments. This turned out not
to be sufficient, so we then added a backup restore process which restores the
build history to the container when it starts up.

This means that the EC2 directory mount is no longer necessary, because the
container gets all of the Jenkins configuration from the combination of the
image and the restored backup.