Increase the health check threshold
Increase the unhealthy check threshold from the default (2) to 5. This
is the number of times that a healthcheck has to fail before ECS deploys
a new container.

Healthchecks currently run every 30 seconds. The previous limit meant that
the Jenkins container would be redeployed if Jenkins didn't response to
the healthcheck within a minute. That meant that we couldn't soft-restart
the Jenkins service, which takes around 90 seconds to restart.

Increasing this threshold means that Jenkins might be down for a few
minutes before ECS tries to resolve the problem, but we don't expect that
to be a big problem because Jenkins is developer-facing and we'll be able
to investigate any problems that arise in practice.