Remove the python3.7 image
When we were using the debian version for the build node, there was no
package for python3.7 so it had to be compiled from source which took a
very long time so I created a base image for the aws to build from. But
I've moved the AWS image to alpine and there's a 3.8 package in there so
this is no longer needed.