Add symlink to launch jar.
The database migration publish job was failing because it couldn't find
the sbt-launch.jar file. I don't know why this doesn't affect the front
end build because those docker files are identical in terms of
installing sbt but there you go.
I've added another symlink to the launch jar and the build is working
now.