Configure Jenkins backups
Use AWS Systems Manager Maintenance Window to set up a daily Jenkins
backup.

This PR moves the Maintenance Windows config from tdr-terraform-backend
into this repo: https://github.com/nationalarchives/tdr-terraform-backend/pull/82

Configuring it in the backend repository caused a problem whenever the
Jenkins EC2 instance was redeployed, because the maintenance window is
linked directly to the EC2 _instance_ ID. This meant that when you
deploy the Jenkins EC2 instance, the backups fail because the maintenance
windows tries to run the backups on the old instance.

Moving the configuration to the tdr-jenkins repo should fix the issue
because we deploy a new EC2 instance by running the tdr-jenkins
Terraform script. This should update link between the maintenance
window and the EC2 instance whenever we deploy a new instance.

Only the Maintenance Window config has been moved into this repo, not the
associated CloudWatch event and monitoring task. This is because the
monitoring task, which should alert us via email and Slack, didn't work.
We need to investigate the problem before moving it to this repo.