Add accounts to the frontend container
The e2e tests need to assume a role in the intg/staging accounts in
order to check the consignment export s3 bucket. The sts sdk uses the
arn for the role which means we need the account number. It's easiest to
get these from the environment variables so I've added them to the
container here.