Filter only running images
This is the command for the jenkins backup which is run on the ec2
instance against the jenkins docker container. The command as was works
if there is only one jenkins container but if jenkins has been restarted
recently, there can be old, stopped containers and trying to run docker
exec on these doesn't work.
So we need to filter out the non-running containers.