Change owner to user 1000 for /var/run/docker*
The jenkins docker container is now running as the jenkins user rather
than root. This user is in the docker group but it's still not able to
run docker commands. I think this is because the /var/run/docker.sock
and /var/run/docker/ are owned by root on the host machine so the docker
container can't access them.
This changes permissions on the docker.sock and docker directory which
allows the jenkins user to access them.