Fetch IP allow list from tdr-configurations instead of SSM param
This changes where the list of IPs which are allowed to access Jenkins
is read from.

Before, we read the list from a parameter in the AWS SSM Parameter Store.
Now, it reads the list from the tdr-configurations repo.

The tdr-configurations repo was already being used for the frontend
load balancer allow list, so this change makes Jenkins consistent
with tdr-terraform-environments.

It also reduces the number of deployment steps. When the list changes,
we can now deploy the changes by running this Terraform project,
without having to deploy the change to the parameter store first.