Update the Dockerfile to create the releases table
There are three changes here.
The JAVA_OPTS change sets the CSP to allow Jenkins to load css from the
userContent directory. Without this, the csp forbids it.

I've installed the quik library which is the templating library we use
to create the html file with the table of release versions. Normally
we'd create a new node image for something like this but it's a lot of
overhead compared to a small change on the main Jenkins node.

The last change is to set the host variable. This is used inside the
html file to link the correct Jenkins instance for the css. I can't find
a better way of doing this.