Add secret for slack release channel webhook.
I'm using a slack webhook to send updates about out of date versions and
we would like to send this to the release channel. We don't currently
have a webhook for the release channel. I've added the parameter into
SSM manually and this will add it into the Jenkins instance.

We don't need this for prod Jenkins as the job will only run on the
integration Jenkins.