Copy from the plugin-updates directory.
We have changed the build process so we build all of the node images
from the nodes directory and use `-f <node_name>/Dockerfile` in the
build command. This copy command was copying the contents of the nodes
directory instead of the plugin-updates directory it was supposed to so
this corrects it.