Account for different location of docker files
The Jenkins parent nodes docker files are located in a different directory to the child node docker files

Child node docker files rely on script file located in their parent directory './docker/nodes'. This requires the docker commands to be run in the parent directory, using the '-f' flag in the command. This is not needed for the Jenkins parent nodes.