Tag Jenkins node images with version number
To avoid having duplicate image tag of "jenkins-prod:latest" when building the Jenkins prod image node on the Jenkins prod instance, use version number as the tag, then retag the pushed image on ECR with the "latest".

Having a duplicate image tag causes Docker to untag the earlier image, ie the image which the Jenkins prod container is built from. This means that Jenkins prod container references the image id instead of the tag. This causes the Jenkins backup script to fail as the script looks for the container using the image tag, which is no longer referenced by the running container.