Install npm from apk
The latest busybox upgrade seems to have broken the npm install which is
why it's a bit risky installing the npm binaries like we have been.

It's safer to install them from apk as this is built for alpine. The
version is 7.22 something which is up to date enough for what we need it
for.